# Conflicts:
#	Engine/platform/windows/acplwin.cpp
#	Engine/platform/windows/media/video/acwavi.cpp
#	Engine/platform/windows/setup/winsetup.cpp

+++ Engine/ac/sprite.cpp
@@ -125,7 +125,7 @@ Bitmap *tmpdbl, *curspr;
 int newwid, newhit;
 void initialize_sprite (int ee) {
 
-    if ((ee < 0) || (ee > spriteset.GetSpriteSlotCount()))
+    if ((ee < 0) || ((size_t)ee > spriteset.GetSpriteSlotCount()))
         quit("initialize_sprite: invalid sprite number");
 
     if ((spriteset[ee] == nullptr) && (ee > 0)) {
